Changes / Fixed bugs in version 2.0:
Biggest changes:

•	New, rewritten tile collision system that is more accurate, faster,
more stable, less bugs, require a lot less code and is less cpu
intensive
•	New, almost rewritten Tile System that is a lot more Object Oriented. 

Changes TileSystem:

•	Tiles are now in their own package
•	All tiles are now Tile objects inside a GameTile object
•	Tiles are now more separated from the GameEngine
•	GameEngine has now only 1 static GameTile object that is used for the
GameTiles. For now a maximum of 1 GameTile object is premited
•	setTileMap() method in GameEngine now adds a whole GameTiles object
instead of the separated map and tileresources array.
•	drawTiles is now called in the GameTiles object instead of the draw
function in the gameEngine
•	getMapWidth() and getMapHeight() moved from GameEngine to GameTile
•	findTilesAt() is moved to MoveableGameObject
•	Setting a starting point for the Game Tiles is not possible anymore.
The function was too unstable to use.
•	DebugMode is not possible anymore with this new tile system

Changes Collision system:

•	Collision calculation with tiles don’t happens anymore at
moveableGameObject. This new collision system has a object called
‘collidingobject’ that has all the collision calculations in it. At
every move the moveablegameobject make the colliding object will
calculate if there is a collision at the current point and the next
moving point. This prevents a fast moving MoveableGameObject to pass
through an tile. When there is a collision, collidingobject will trigger
CollisionOccured (that is already implemented in MoveableGameObject) of
a ICollision object.
•	Collision will now works with CollisionOccured(List<Tile> tiles)
•	Methods that are related to tile collision are rewritten (like
moveUpToTileSide())
•	All handle….TileCollision() methods are deleted . Now there is a
method called ‘getCollisionSide()’ that can be used for this
functionality

Other bugfixes / changes:

•	Collision with Gameobject with another (moveable)GameObject works
correct now
•	getCollidedObject() method is deleted because it added nothing really
usefull
•	Animated sprite only showed a small version of the sprite (see issue 1
Github)
•	Sprite now delays loading when it notices that the surface has not
created yet
•	Initialize() in now called when the GameLoop thread starts (directly
before the first update() call). This ensures that the surface has been
created before intilialize() is called.
•	startGame() is deprecated. The game starts automatically. startGame()
was first used to initialize the touchinput. Initializing of the
touchinput is now done at the start of the gameLoop thread.
•	Maximum fps decreased to 30 FPS for more stable game speed across
devices. Maximum fps can always be changed using the GameLoop.MAX_FPS
flag.
•	2 new types of addGameObject added. One for if you only want to add a
gameObject, and one when you want to add a gameObject and a layer
position. Previously you had to add a startPosition, that’s not required
anymore.
•	Player automatically jumps to its start position now.
•	setStartPostion() now not automatically jumps to the start position.
•	Menu deleted because it’s no longer supported by google
•	outsideWorld has now an parameter to check if a object has leaved the
game world horizontal or vertical
•	getAngle() moved from MoveableGameObject to GameObject
•	CardinalDirection class is deleted because it had unusable and
inconsequent information
•	Flag GameView.surfaceloaded added that switch to true when the surface
has been loaded
•	Lot of unnecessary code deleted in the onDraw() method of GameView
•	Tiles are now drawn behind GameObjects
•	Update() in moveableGameObject is cleaned up and updatePlayer() is
deleted

Pfew, that was a whole list, and I think I haven’t even say every change
;) . For more than 40+ changes, this beta version 2 is a nice follow up
to the previous version.

Hopefully you will enjoy it. If you find any bugs, have suggestions or
whatsoever, don’t doubt to write a issue, suggestion , or fork the
project to make your own implementation.

--Bas